renamed observer.py into utils.py; changed imports for single_objectives
