+++ docs/notebooks/multi_objective_ehvi.pct.py
@@ -33,7 +33,7 @@ tf.random.set_seed(1793)
 
 # %%
 vlmop2 = VLMOP2().objective()
-observer = trieste.objectives.observer.mk_observer(vlmop2)
+observer = trieste.objectives.utils.mk_observer(vlmop2)
 
 # %%
 mins = [-2, -2]
