+++ trieste/objectives/utils.py
